requirements.txt
alabaster
Babel
cairocffi
CairoSVG
certifi
cffi
chardet
# click
colorama
cryptography
cssselect2
defusedxml
diff-match-patch
Django
django-cors-headers
django-filter
django-import-export
django-js-asset
django-mptt
django-simple-email-confirmation
django-wkhtmltopdf
djangorestframework
docopt
docutils
et-xmlfile
# feedparser --no-build-isolation
# fpdf
gitsome
gunicorn
html5lib
idna
imagesize
jdcal
Jinja2
jwcrypto
MarkupPy
MarkupSafe
# mysqlclient==1.4.6
numpy
numpydoc
oauthlib
odfpy
openpyxl
packaging
pandas
Pillow
ply
prompt-toolkit
public
pycparser
Pygments
PyJWT
pyparsing
Pyphen
python-dateutil
python3-openid
pytz
PyYAML
query-string
reportlab
requests
requests-oauthlib
six
snowballstemmer
social-auth-app-django
social-auth-core
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
sqlparse
tablib
tinycss2
twilio
unicodecsv
uritemplate
uritemplate.py
urllib3
wcwidth
WeasyPrint
webencodings
xlrd
XlsxWriter
xlwt
